requirements.txt
torch>=2.2.0
torchvision>=0.17.0
hydra-core>=1.3.0
omegaconf>=2.3.0
wandb>=0.16.0
webdataset>=0.2.86
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
hydra-submitit-launcher>=1.2.0
submitit>=1.5.0
h5py>=3.9.0
